Refaktor: Pakkestruktur domene og separate filer for domeneklasser
